pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "pdm-dotenv"
description = "A pdm plugin that automatically loads .env files"
authors = [{ name = "Zane Dufour", email = "zane@znd4.me" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pdm>=2.5.2", "python-dotenv>=1.0.0"]
dynamic = ["version"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
]

[project.entry-points.pdm]
dotenv = "pdm_dotenv:plugin"

[project.urls]
Repository = "https://github.com/zdog234/pdm-dotenv"
Homepage = "https://github.com/zdog234/pdm-dotenv"

[tool.pdm.version]
source = "scm"

[tool.pdm.build]
package-dir = "src"

[tool.pdm.dev-dependencies]
test = ["pytest>=6.1", "pdm[pytest]>=2.5.2"]
dev = [
    "towncrier>=19.2",
    "parver>=0.3",
    "pre-commit>=3.2.2",
    "nox>=2022.11.21",
    "black>=23.3.0",
    "pdm[pytest]>=2.5.2",
]

doc = ["mkdocs>=1.1", "mkdocs-material>=6.2"]

[tool.pdm.scripts]
post_install = "pre-commit install"
release = "python -m tasks.release"
test = "pytest tests/"
doc = { shell = "cd docs && mkdocs serve", help = "Start the dev server for doc preview" }
lint = "pre-commit run --all-files"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "E",   # pycodestyle
    "F",   # pyflakes
    "PGH", # pygrep-hooks
    "RUF", # ruff
    "W",   # pycodestyle
    "YTT", # flake8-2020
]
extend-ignore = ["B018", "B019"]
src = ["src"]
extend-exclude = ["tests/fixtures"]
target-version = "py37"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.isort]
known-first-party = ["pdm_dotenv"]

[tool.mypy]
follow_imports = "silent"
ignore_missing_imports = true
disallow_incomplete_defs = true
check_untyped_defs = false
disallow_untyped_decorators = true
namespace_packages = true
mypy_path = "src"
explicit_package_bases = true

[tool.towncrier]
package = "pdm_dotenv"
filename = "CHANGELOG.md"
issue_format = "[#{issue}](https://github.com/zdog234/pdm-dotenv/issues/{issue})"
directory = "news/"
start_string = "<!-- insertion marker -->"
title_format = "[v{version}](https://github.com/zdog234/pdm-dotenv/releases/tag/{version}) ({project_date})"
template = "news/towncrier_template.md"
underlines = "-~^"

[[tool.towncrier.type]]
directory = "feature"
name = "Features & Improvements"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bug Fixes"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Improved Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "dep"
name = "Dependencies"
showcontent = true

[[tool.towncrier.type]]
directory = "removal"
name = "Removals and Deprecations"
showcontent = true

[[tool.towncrier.type]]
directory = "misc"
name = "Miscellany"
showcontent = true

[tool.isort]
profile = "black"
atomic = true
skip_glob = ["*/setup.py"]
filter_files = true
known_first_party = ["pdm_dotenv"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
